+++ packages/itwin/breakdown-trees/src/test/functionalityProviders/ToggledTopFitViewFunctionalityProvider.test.ts
@@ -26,7 +26,7 @@ describe("ToggledTopFitViewFunctionalityProvider", () => {
     await IModelApp.localization.registerNamespace("BreakdownTrees");
 
     const ifcWallNodeKey = FunctionalityProviderTestUtils.createClassNodeKey([], [FunctionalityProviderTestUtils.createECInstanceKey(MockClassNames.IfcWall, "0x3")]);
-    dataProviderMock.setup((x) => x.getNodeKey(moq.It.isObjectWith<TreeNodeItem>({ id: MockStrings.IfcWallNode }))).returns((_item: TreeNodeItem): ECInstancesNodeKey => ifcWallNodeKey);
+    dataProviderMock.setup((x) => x.getNodeKey(moq.It.isObjectWith<TreeNodeItem>({ id: MockStrings.IfcWallNode }))).returns((_item: TreeNodeItem): ECInstancesNodeKey => ifcWallNodeKey); // eslint-disable-line deprecation/deprecation
 
     handleTopViewSpy = sinon.stub(ToggledTopFitViewFunctionalityProvider.prototype, "handleTopView" as any);