+++ packages/itwin/grouping-mapping-widget/src/widget/components/GroupAction.tsx
@@ -2,13 +2,6 @@
 * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/
-import type { IModelConnection } from "@itwin/core-frontend";
-import type {
-  ISelectionProvider,
-  SelectionChangeEventArgs,
-} from "@itwin/presentation-frontend";
-import { Presentation } from "@itwin/presentation-frontend";
-import { useActiveIModelConnection } from "@itwin/appui-react";
 import type {
   SelectOption,
 } from "@itwin/itwinui-react";
@@ -23,25 +16,17 @@ import {
 } from "./utils";
 import "./GroupAction.scss";
 import useValidator from "../hooks/useValidator";
-import {
-  clearEmphasizedElements,
-  clearOverriddenElements,
-  transparentOverriddenElements,
-  visualizeElementsByQuery,
-  zoomToElements,
-} from "./viewerUtils";
 import { useGroupingMappingApiConfig } from "./context/GroupingApiConfigContext";
 import { useMappingClient } from "./context/MappingClientContext";
 import { useGroupingMappingCustomUI } from "./context/GroupingMappingCustomUIContext";
 import type { GroupingCustomUI } from "./customUI/GroupingMappingCustomUI";
 import { GroupingMappingCustomUIType } from "./customUI/GroupingMappingCustomUI";
 import type { Group } from "@itwin/insights-client";
-import { useGroupHilitedElementsContext } from "./context/GroupHilitedElementsContext";
-import { visualizeGroupColors } from "./groupsHelpers";
 import { QueryBuilderStep } from "./QueryBuilderStep";
 import { GroupDetailsStep } from "./GroupDetailsStep";
 import { QueryBuilderActionPanel } from "./QueryBuilderActionPanel";
 import { GroupDetailsActionPanel } from "./GroupDetailsActionPanel";
+import { useVisualization } from "../hooks/useVisualization";
 
 const defaultDisplayStrings = {
   groupDetails: "Group Details",
@@ -55,6 +40,7 @@ enum GroupActionStep {
 
 export interface GroupActionProps {
   mappingId: string;
+  shouldVisualize: boolean;
   group?: Group;
   queryGenerationType: string;
   onSaveSuccess: () => void;
@@ -63,9 +49,10 @@ export interface GroupActionProps {
 }
 
 export const GroupAction = (props: GroupActionProps) => {
-  const iModelConnection = useActiveIModelConnection() as IModelConnection;
-  const { showGroupColor, groups, hiddenGroupsIds, hilitedElementsQueryCache } = useGroupHilitedElementsContext();
-  const { getAccessToken, iModelId } = useGroupingMappingApiConfig();
+  const { getAccessToken, iModelId, iModelConnection } = useGroupingMappingApiConfig();
+  if (!iModelConnection) {
+    throw new Error("This component requires an active iModelConnection.");
+  }
   const mappingClient = useMappingClient();
   const groupUIs: GroupingCustomUI[] = useGroupingMappingCustomUI().customUIs
     .filter((p) => p.type === GroupingMappingCustomUIType.Grouping) as GroupingCustomUI[];
@@ -74,21 +61,29 @@ export const GroupAction = (props: GroupActionProps) => {
     description: props.group?.description ?? "",
   });
   const [query, setQuery] = useState<string>("");
-  const [simpleSelectionQuery, setSimpleSelectionQuery] = useState<string>("");
+  const [queryRowCount, setQueryRowCount] = useState<number>(0);
+
   const [validator, setShowValidationMessage] = useValidator();
   const [isLoading, setIsLoading] = useState<boolean>(false);
-  const [isRendering, setIsRendering] = useState<boolean>(false);
+
   const [queryGenerationType, setQueryGenerationType] = useState(
     props.queryGenerationType,
   );
+  const {
+    isRendering,
+    setIsRendering,
+    simpleSelectionQuery,
+    setSimpleSelectionQuery,
+    clearPresentationSelection,
+    resetView,
+  } = useVisualization(
+    props.shouldVisualize,
+    iModelConnection,
+    query,
+    queryGenerationType
+  );
   const isUpdating = isLoading || isRendering;
   const [currentStep, setCurrentStep] = React.useState(GroupActionStep.QueryBuilder);
-
-  useEffect(() => {
-    if (!iModelConnection) {
-      throw new Error("This component requires an active iModelConnection.");
-    }
-  }, [iModelConnection]);
 
   const displayStrings = React.useMemo(
     () => ({ ...defaultDisplayStrings, ...props.displayStrings }),
@@ -97,68 +92,53 @@ export const GroupAction = (props: GroupActionProps) => {
 
   useEffect(() => setQueryGenerationType(props.queryGenerationType), [props.queryGenerationType]);
 
-  const resetView = useCallback(async () => {
-    if (showGroupColor) {
-      await visualizeGroupColors(iModelConnection, groups, groups, hiddenGroupsIds, hilitedElementsQueryCache);
-    } else {
-      clearOverriddenElements();
-    }
-    clearEmphasizedElements();
-  }, [groups, hiddenGroupsIds, hilitedElementsQueryCache, iModelConnection, showGroupColor]);
-
   useEffect(() => {
-    const removeListener = Presentation.selection.selectionChange.addListener(
-      async (
-        evt: SelectionChangeEventArgs,
-        selectionProvider: ISelectionProvider,
-      ) => {
-        if (queryGenerationType === "Selection") {
-          const selection = selectionProvider.getSelection(
-            evt.imodel,
-            evt.level,
-          );
-          const query = selection.instanceKeys.size > 0
-            ? `SELECT ECInstanceId FROM ${selection.instanceKeys.keys().next().value}`
-            : "";
-          setSimpleSelectionQuery(query);
-        }
-      },
-    );
-    return () => {
-      removeListener();
-    };
-  }, [iModelConnection, queryGenerationType]);
-
-  useEffect(() => {
-    const reemphasize = async () => {
+    const fetchQueryRowCount = async () => {
       try {
         if (!query || query === "") {
+          setQueryRowCount(0);
           return;
         }
-        setIsRendering(true);
-        transparentOverriddenElements();
-        const resolvedHiliteIds = await visualizeElementsByQuery(
-          query,
-          "red",
-          iModelConnection,
-        );
-        await zoomToElements(resolvedHiliteIds);
+        setIsLoading(true);
+        const result = await iModelConnection.queryRowCount(query);
+        setQueryRowCount(result);
       } catch {
-        toaster.negative("Sorry, we have failed to generate a valid query. 😔");
+        toaster.negative("Query failed to resolve.");
       } finally {
-        setIsRendering(false);
+        setIsLoading(false);
       }
     };
 
-    void reemphasize();
-  }, [iModelConnection, query]);
+    void fetchQueryRowCount();
+  }, [iModelConnection, query, setIsRendering]);
 
-  useEffect(() => {
-    Presentation.selection.clearSelection(
-      "GroupingMappingWidget",
-      iModelConnection,
-    );
-  }, [iModelConnection]);
+  const isBlockingActions = !(
+    details.groupName &&
+    (query || simpleSelectionQuery) &&
+    !isRendering &&
+    !isLoading
+  );
+
+  const getOptions = useMemo(
+    (): SelectOption<string>[] =>
+      groupUIs.map((ui) => ({
+        label: ui.displayLabel,
+        value: ui.name,
+        icon: ui.icon,
+      })),
+    [groupUIs]
+  );
+
+  const onChange = useCallback(
+    async (value: string) => {
+      setQueryGenerationType(value);
+      clearPresentationSelection();
+      setQuery("");
+      setSimpleSelectionQuery("");
+      await resetView();
+    },
+    [clearPresentationSelection, resetView, setSimpleSelectionQuery]
+  );
 
   const save = useCallback(async () => {
     if (!validator.allValid()) {
@@ -188,10 +168,7 @@ export const GroupAction = (props: GroupActionProps) => {
             query: currentQuery,
           },
         );
-      Presentation.selection.clearSelection(
-        "GroupingMappingWidget",
-        iModelConnection,
-      );
+      clearPresentationSelection();
       setDetails({
         groupName: props.group?.groupName ?? "",
         description: props.group?.description ?? "",
@@ -204,38 +181,7 @@ export const GroupAction = (props: GroupActionProps) => {
     } finally {
       setIsLoading(false);
     }
-  }, [validator, setShowValidationMessage, query, simpleSelectionQuery, getAccessToken, props, mappingClient, iModelId, details, iModelConnection]);
-
-  const isBlockingActions = !(
-    details.groupName &&
-    (query || simpleSelectionQuery) &&
-    !isRendering &&
-    !isLoading
-  );
-
-  const getOptions = useMemo(
-    (): SelectOption<string>[] =>
-      groupUIs.map((ui) => ({
-        label: ui.displayLabel,
-        value: ui.name,
-        icon: ui.icon,
-      })),
-    [groupUIs]
-  );
-
-  const onChange = useCallback(
-    async (value: string) => {
-      setQueryGenerationType(value);
-      Presentation.selection.clearSelection(
-        "GroupingMappingWidget",
-        iModelConnection,
-      );
-      setQuery("");
-      setSimpleSelectionQuery("");
-      await resetView();
-    },
-    [iModelConnection, resetView]
-  );
+  }, [validator, setShowValidationMessage, query, simpleSelectionQuery, getAccessToken, props, mappingClient, iModelId, details, clearPresentationSelection]);
 
   const isQueryBuilderStep = currentStep === GroupActionStep.QueryBuilder;
   const isGroupDetailsStep = currentStep === GroupActionStep.GroupDetails;
@@ -243,6 +189,7 @@ export const GroupAction = (props: GroupActionProps) => {
     <>
       <div className="gmw-group-add-modify-container">
         <QueryBuilderStep
+          queryRowCount={queryRowCount}
           isHidden={!isQueryBuilderStep}
           queryGenerationType={queryGenerationType}
           groupUIs={groupUIs}
@@ -252,6 +199,7 @@ export const GroupAction = (props: GroupActionProps) => {
           onChange={onChange}
           getOptions={getOptions}
           displayStrings={{ ...displayStrings }}
+          group={props.group}
         />
         {isGroupDetailsStep && <GroupDetailsStep
           details={details}
@@ -277,11 +225,8 @@ export const GroupAction = (props: GroupActionProps) => {
         {props.onClickCancel && <Button
           type='button'
           id='cancel'
-          onClick={async () => {
-            Presentation.selection.clearSelection(
-              "GroupingMappingWidget",
-              iModelConnection,
-            );
+          onClick={() => {
+            clearPresentationSelection();
             props.onClickCancel && props.onClickCancel();
           }}
         >